requirements.txt
apache-airflow[postgres,amazon]
psycopg2-binary
boto3
